% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\name{runMathQuest}
\alias{runMathQuest}
\title{runMathQuest}
\usage{
runMathQuest(name, question_count)
}
\arguments{
\item{name}{string. A string containing a name to personalize the UI input.}

\item{question_count}{numeric. Number of questions the user wants the quiz to contain.}
}
\description{
Math Quest is a fun and playful quiz app that allows elementary school children (grade 4 through grade 8) to practice different parts of math, such as addition, subtraction, multiplication, and division. The app generates math questions adapted to the selected school grade level. The application incorporates gamification and learning strategies to promote and make learning more fun for students and optimize their enjoyment to achieve the best results.
}
\examples{
# Run Math Quest quiz with 15 questions
runMathQuest(name = "Bruny", question_count = 15)

# Run Math Quest quiz with 25 questions and different name
runMathQuest(name = "Ruben", question_count = 25)


}
